Allows renaming ids without losing the translation
In order to work, you must provide a `old` key to the defineMessages,
that will be used to match the old id and keep the translation with the
new id.

This solution relies at changing the babel-plugin-react-intl, in
specifically, the DESCRIPTOR_PROPS and the storeMessage function,
to allow the `old` property be extracted.

There's must be a better solution without relying at hard coding it.

e.g.:

const t = defineMessages({
  name: {
    id: 'hello_name',
    defaultMessage: 'Hello',
  }
})

nl.json
{
  "hello_name": "Uw naam"
}

Then, add a `old` property to the defineMessages.

const t = defineMessages({
  name: {
    id: 'your_name',
    old: 'hello_name',
    defaultMessage: 'Hello',
  }
})

Renamed keys:
  hello_name -> your_name: Uw naam

nl.json
{
  "your_name": "Uw naam"
}